requirements.txt
flake8==3.7.1
pycodestyle==2.5.0
pyyaml